requirements.txt
torch
pillow
numpy
ipywidgets
scikit-image
matplotlib